requirements.txt
fastapi[standard]
psycopg[binary,pool]
SQLAlchemy
passlib[bcrypt]
python-jose[crytpography]
pydantic-settings
alembic